Create util directory. Create shopping.js module, begin populating it with all relevant shopping related functions
