+++ src/components/ShoppingCartLimitModal/ShoppingCartLimitModal.css
@@ -46,9 +46,9 @@
 .limit-inner-wrapper .limit-btn-wrapper {
   display: flex;
   flex-direction: row;
-  gap: 8px;
   justify-content: flex-end;
   align-items: center;
+  gap: 8px;
   margin-bottom: 25px;
 }
